+++ tests/e2e/env/utils/app-name.js
@@ -5,7 +5,7 @@ const path = require( 'path' );
 /**
  * Internal dependencies
  */
-const getTestConfig = require( './test-config' );
+const { getTestConfig } = require( './test-config' );
 const getAppRoot = require( './app-root' );
 
 const getAppName = () => {
@@ -13,8 +13,15 @@ const getAppName = () => {
 	if ( testConfig.appName ) {
 		return testConfig.appName;
 	}
-	const appRoot = getAppRoot();
-	return path.basename( appRoot );
+	return getAppBase();
 };
 
-module.exports = getAppName;
+const getAppBase = () => {
+	const appRoot = getAppRoot();
+	return path.basename( appRoot );
+}
+
+module.exports = {
+	getAppName,
+	getAppBase,
+};