+++ tests/e2e/core-tests/CHANGELOG.md
@@ -1,10 +1,19 @@
 # Unreleased
 
+# 0.1.1
+
 ## Added
+
+- Registered Shopper Checkout tests
 - Merchant Order Status Filter tests
 - Merchant Order Refund tests
 - Merchant Apply Coupon tests
 - Added new config variable for Simple Product price to `tests/e2e/env/config/default.json`. Defaults to 9.99
+
+- Shopper Checkout Apply Coupon
+
+
+- Shopper Cart Apply Coupon
 
 ## Fixed